backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
transformers
pillow
numpy
opencv-python
grad-cam
